This code successfully tokenizes and saves a dataset, and will check if a dataset has been tokenized in the fit method, using that if available else it processes the data one piece at a time. (Although I'm not 100% sure that it is processing correctly, because it outputs such a simple tensor. That is something worth reviewing).

+++ src/mirror/datasets/util.py
@@ -51,17 +51,12 @@ def load_hf_from_cache_or_download(
 
     return ds
 
-
-def tokenized_data_path(dataset_id: str) -> Path:
-    return Path(f'/home/{os.environ['USER']}/nobackup/autodelete/mirror_data/tokenized_data/{dataset_id}_tokenized/')
-
-
 def load_tokenized_from_cache(
-        hf_dataset_path: str,
-        hf_dataset_name: str | None = None,
-        # process: Callable[[Dataset | DatasetDict], Dataset | DatasetDict] | None = None,
-        tokenizer_function: Callable = None,
+        dataset: Dataset | None,
+        dataset_id: str,
+        tokenizer_function: Callable | None = None,
         reset_cache: bool = False,
+        preprocess: bool = False,
 ) -> Dataset | DatasetDict:
     """
     The first time this is called with a particular dataset, it will tokenize
@@ -73,27 +68,25 @@ def load_tokenized_from_cache(
     This means if your process function should not rely on information that might change
     independently of the path/name (such as split).
     """
-    dataset_path = datasets_path / hf_dataset_path / hf_dataset_name \
-        if hf_dataset_name else datasets_path / hf_dataset_path
-
-    if reset_cache:
-        shutil.rmtree(dataset_path)
+    dataset_path = Path(mirror_data_path / f'tokenized_data/{dataset_id}')
 
     is_cached = os.path.exists(dataset_path)
 
-    if is_cached:
-        ds = load_from_disk(dataset_path)
-    else:
-        assert_can_download(hf_dataset_path)
-        ds = load_dataset(
-            hf_dataset_path,
-            hf_dataset_name,
-            cache_dir=str(datasets_path / hf_dataset_path)
-        )
-        assert isinstance(ds, Dataset) or isinstance(ds, DatasetDict)
-        if tokenizer_function:
-            ds = ds.map(tokenizer_function)
-        
-        ds.save_to_disk(dataset_path)
+    if reset_cache and is_cached:
+        print("Resetting cache.")
+        shutil.rmtree(dataset_path)
+        is_cached = False
 
-    return ds
+    if is_cached:
+        dataset = load_from_disk(dataset_path)
+        print("Dataset already tokenized & loaded from disk.")
+    else:
+        print("Dataset is not yet cached.")
+        if preprocess and dataset:
+            dataset = dataset.map(tokenizer_function)
+            print("Dataset tokenized.")
+            dataset.save_to_disk(dataset_path)
+            print("Dataset saved to disk.")
+            is_cached = True
+
+    return dataset, is_cached
